This is a combination of 2 commits.
feat: allow FE features in browser

feat: allow `generateViewScript` and `isDropInStatements` FE feature in browser

feat: allow `generateContainerScript` FE feature in browser; eliminated circular dependencies
